refact: wrapPromise utils로 리팩토링
